backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
python-docx==1.1.2
pdfplumber==0.11.1
PyMuPDF==1.24.3
openpyxl==3.1.2
google-generativeai==0.7.2
langdetect==1.0.9
aiofiles==23.2.1
httpx==0.27.0
python-dotenv==1.0.1
